#372 Extend notification target messages with more fields and add pipeline notification target interface.
