+++ pramen/core/src/test/scala/za/co/absa/pramen/core/mocks/TaskResultFactory.scala
@@ -16,8 +16,8 @@
 
 package za.co.absa.pramen.core.mocks
 
+import za.co.absa.pramen.api.SchemaDifference
 import za.co.absa.pramen.core.mocks.job.JobSpy
-import za.co.absa.pramen.core.notify.pipeline.SchemaDifference
 import za.co.absa.pramen.core.pipeline.{DependencyWarning, Job, TaskRunReason}
 import za.co.absa.pramen.core.runner.task.{NotificationFailure, RunInfo, RunStatus, TaskResult}
